move digi-id parsing methods to `pybes3.digi_id`
