+++ docs/api/pybes3.detectors.md
@@ -1,31 +1,9 @@
 # pybes3.detectors
 
-## Parsing methods
-::: pybes3.detectors.parse_mdc_gid
----
-::: pybes3.detectors.parse_mdc_digi_id
----
-::: pybes3.detectors.parse_tof_digi_id
----
-::: pybes3.detectors.parse_emc_digi_id
----
-::: pybes3.detectors.parse_muc_digi_id
----
-::: pybes3.detectors.parse_cgem_digi_id
----
-
 ## MDC
-::: pybes3.detectors.get_mdc_wire_position
----
 ::: pybes3.detectors.get_mdc_gid
 ---
-::: pybes3.detectors.mdc_gid_to_layer
----
-::: pybes3.detectors.mdc_gid_to_wire
----
-::: pybes3.detectors.mdc_gid_to_is_stereo
----
-::: pybes3.detectors.mdc_layer_to_is_stereo
+::: pybes3.detectors.get_mdc_wire_position
 ---
 ::: pybes3.detectors.mdc_gid_to_east_x
 ---
@@ -33,15 +11,31 @@
 ---
 ::: pybes3.detectors.mdc_gid_to_east_z
 ---
+::: pybes3.detectors.mdc_gid_to_is_stereo
+---
+::: pybes3.detectors.mdc_gid_to_layer
+---
+::: pybes3.detectors.mdc_gid_to_stereo
+---
+::: pybes3.detectors.mdc_gid_to_superlayer
+---
 ::: pybes3.detectors.mdc_gid_to_west_x
 ---
 ::: pybes3.detectors.mdc_gid_to_west_y
 ---
 ::: pybes3.detectors.mdc_gid_to_west_z
 ---
+::: pybes3.detectors.mdc_gid_to_wire
+---
 ::: pybes3.detectors.mdc_gid_z_to_x
 ---
 ::: pybes3.detectors.mdc_gid_z_to_y
+---
+::: pybes3.detectors.mdc_layer_to_is_stereo
+---
+::: pybes3.detectors.mdc_layer_to_superlayer
+---
+::: pybes3.detectors.parse_mdc_gid
 ---
 
 ## EMC
@@ -86,4 +80,6 @@
 ::: pybes3.detectors.emc_gid_to_point_z
 ---
 ::: pybes3.detectors.emc_gid_to_theta
+---
+::: pybes3.detectors.parse_emc_gid
 ---
