+++ docs/api/pybes3.digi_id.md
@@ -1,3 +1,98 @@
 # pybes3.digi_id
 
-::: pybes3.digi_id
+## Parsing methods
+
+These methods parse the given ID and return the corresponding components. They are useful for retrieving the components of an ID when you only have the ID itself.
+
+::: pybes3.digi_id.parse_mdc_digi
+---
+::: pybes3.digi_id.parse_tof_digi_id
+---
+::: pybes3.digi_id.parse_muc_digi_id
+---
+::: pybes3.digi_id.parse_cgem_digi_id
+---
+
+## MDC
+::: pybes3.digi_id.check_mdc_id
+---
+::: pybes3.digi_id.mdc_id_to_wire
+---
+::: pybes3.digi_id.mdc_id_to_layer
+---
+::: pybes3.digi_id.mdc_id_to_is_stereo
+---
+::: pybes3.digi_id.get_mdc_digi_id
+---
+::: pybes3.digi_id.parse_mdc_digi_id
+---
+::: pybes3.digi_id.parse_emc_digi_id
+---
+::: pybes3.digi_id.parse_emc_digi
+---
+
+## TOF
+::: pybes3.digi_id.check_tof_id
+---
+::: pybes3.digi_id.tof_id_to_part
+---
+::: pybes3.digi_id.tof_id_to_end
+---
+::: pybes3.digi_id._tof_id_to_layer_or_module_1
+---
+::: pybes3.digi_id._tof_id_to_layer_or_module_2
+---
+::: pybes3.digi_id.tof_id_to_layer_or_module
+---
+::: pybes3.digi_id._tof_id_to_phi_or_strip_1
+---
+::: pybes3.digi_id._tof_id_to_phi_or_strip_2
+---
+::: pybes3.digi_id.tof_id_to_phi_or_strip
+---
+::: pybes3.digi_id.get_tof_digi_id
+---
+
+## EMC
+::: pybes3.digi_id.check_emc_id
+---
+::: pybes3.digi_id.emc_id_to_module
+---
+::: pybes3.digi_id.emc_id_to_theta
+---
+::: pybes3.digi_id.emc_id_to_phi
+---
+::: pybes3.digi_id.get_emc_digi_id
+---
+
+## MUC
+::: pybes3.digi_id.check_muc_id
+---
+::: pybes3.digi_id.muc_id_to_part
+---
+::: pybes3.digi_id.muc_id_to_segment
+---
+::: pybes3.digi_id.muc_id_to_layer
+---
+::: pybes3.digi_id.muc_id_to_channel
+---
+::: pybes3.digi_id.get_muc_digi_id
+---
+::: pybes3.digi_id.muc_id_to_gap
+---
+::: pybes3.digi_id.muc_id_to_strip
+---
+
+## CGEM
+::: pybes3.digi_id.check_cgem_id
+---
+::: pybes3.digi_id.cgem_id_to_layer
+---
+::: pybes3.digi_id.cgem_id_to_sheet
+---
+::: pybes3.digi_id.cgem_id_to_strip
+---
+::: pybes3.digi_id.cgem_id_to_is_x_strip
+---
+::: pybes3.digi_id.get_cgem_digi_id
+---
