+++ src/pybes3/__init__.py
@@ -47,13 +47,16 @@ from pybes3.detectors import (
     mdc_gid_z_to_y,
     mdc_layer_to_is_stereo,
     mdc_layer_to_superlayer,
+    parse_emc_gid,
+    parse_mdc_gid,
+)
+
+from .digi_id import (
     parse_cgem_digi_id,
     parse_emc_digi,
     parse_emc_digi_id,
-    parse_emc_gid,
     parse_mdc_digi,
     parse_mdc_digi_id,
-    parse_mdc_gid,
     parse_muc_digi_id,
     parse_tof_digi_id,
 )
@@ -99,6 +102,7 @@ __all__ = [
     "emc_gid_to_theta",
     "get_emc_crystal_position",
     "get_emc_gid",
+    "parse_emc_gid",
     "get_mdc_gid",
     "get_mdc_wire_position",
     "mdc_gid_to_east_x",
@@ -116,13 +120,13 @@ __all__ = [
     "mdc_gid_z_to_y",
     "mdc_layer_to_is_stereo",
     "mdc_layer_to_superlayer",
+    "parse_mdc_gid",
+    # digi_id
     "parse_cgem_digi_id",
     "parse_emc_digi",
     "parse_emc_digi_id",
-    "parse_emc_gid",
     "parse_mdc_digi",
     "parse_mdc_digi_id",
-    "parse_mdc_gid",
     "parse_muc_digi_id",
     "parse_tof_digi_id",
     # tracks
